Move swing related components to swing package
