+++ src/main/java/org/bric/gui/swing/ArrayListTransferHandler.java
@@ -2,7 +2,7 @@
  * To change this template, choose Tools | Templates
  * and open the template in the editor.
  */
-package org.bric.utils;
+package org.bric.gui.swing;
 
 import org.bric.core.model.ImportedImage;
