+++ src/main/java/org/bric/gui/swing/ShadowLabel.java
@@ -2,7 +2,7 @@
  * To change this template, choose Tools | Templates
  * and open the template in the editor.
  */
-package org.bric.gui.watermark;
+package org.bric.gui.swing;
 
 /**
  *
